Fix: Рефакторит главную страницу
