+++ src/components/main-container/index.ts
@@ -1,0 +1,2 @@
+import MainContainer from './main-container';
+export default MainContainer;
